alacritty pretty

+++ system/.zshrc
@@ -108,4 +108,7 @@ fpath+=${ZDOTDIR:-~}/.zsh_functions
 
 # Move Alacritty config to ~/.config
 mkdir -p ~/.config/alacritty
-ln -s $HOME/.dotfiles/system/alacritty.toml ~/.config/alacritty/alacritty.toml
+# ignore if file already exists
+if [ ! -f ~/.config/alacritty/alacritty.toml ]; then
+	ln -s $HOME/.dotfiles/system/alacritty.toml ~/.config/alacritty/alacritty.toml
+fi
